+++ blocklets/image-bin/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.10.13 (September 18, 2023)
+
+- feat: adjust folder logic in Image Bin
+
 ## 0.10.12 (September 16, 2023)
 
 - chore: bump deps to latest